+++ packages/adhere-ui-currency-symbol/es/index.d.ts
@@ -1,0 +1,2 @@
+import CurrencySymbol from './CurrencySymbol';
+export default CurrencySymbol;