+++ packages/adhere-website/src/components/ui/currencysymbol/examples/p1.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import { CurrencySymbol } from '@baifendian/adhere';
+
+export default () => {
+  return <CurrencySymbol amount={999} />;
+};